Abstract DroneDetector class out to ease testing
